+++ R/get_predicted.R
@@ -35,9 +35,16 @@
 #' * `"prediction"` also gives an output on the response scale, but this time
 #'   associated with a prediction interval (PI), which is larger than a confidence
 #'   interval (though it mostly make sense for linear models).
-#' * `"classification"` only differs from `"prediction"` for binomial models
-#'   where it additionally transforms the predictions into the original response's
-#'   type (for instance, to a factor).
+#' * `"classification"` is relevant only for binomial, ordinal or mixture models.
+#'   - For binomial models, `predict = "classification"` will additionally
+#'     transform the predictions into the original response's type (for
+#'     instance, to a factor).
+#'   - For ordinal models (e.g., classes `clm` or `multinom`), gives the
+#'     predicted response class membership, defined as highest probability
+#'     prediction.
+#'   - For finite mixture models (currently only family [`brms::mixture()`] from
+#'     package *brms*), returns a vector of predicted class membership (similar
+#'     as for ordinal models).
 #' * Other strings are passed directly to the `type` argument of the `predict()`
 #'   method supplied by the modelling package.
 #' * Specifically for models of class `brmsfit` (package *brms*), the `predict`
@@ -49,7 +56,7 @@
 #'   by the modelling package. Note that this might result in conflicts with
 #'   multiple matching `type` arguments - thus, the recommendation is to use the
 #'   `predict` argument for those values.
-#' * Notes: You can see the 4 options for predictions as on a gradient from
+#' * Notes: You can see the four options for predictions as on a gradient from
 #'   "close to the model" to "close to the response data": "link", "expectation",
 #'   "prediction", "classification". The `predict` argument modulates two things:
 #'   the scale of the output and the type of certainty interval. Read more about
@@ -138,10 +145,12 @@
 #' and no transformation is applied. For instance, for a logistic regression
 #' model, the response scale corresponds to the predicted probabilities, whereas
 #' the link-scale makes predictions of log-odds (probabilities on the logit
-#' scale). Note that when users select `predict="classification"` in binomial
+#' scale). Note that when users select `predict = "classification"` in binomial
 #' models, the `get_predicted()` function will first calculate predictions as if
-#' the user had selected `predict="expectation"`. Then, it will round the
-#' responses in order to return the most likely outcome.
+#' the user had selected `predict = "expectation"`. Then, it will round the
+#' responses in order to return the most likely outcome. For ordinal or mixture
+#' models, it returns the predicted class membership, based on the highest
+#' probability of classification.
 #'
 #' @section Heteroscedasticity consistent standard errors: The arguments `vcov`
 #' and `vcov_args` can be used to calculate robust standard errors for