+++ R/utils_model_info.R
@@ -1,15 +1,15 @@
-.make_family <- function(x,
-                         fitfam = "gaussian",
-                         zero.inf = FALSE,
-                         hurdle = FALSE,
-                         logit.link = FALSE,
-                         multi.var = FALSE,
-                         link.fun = "identity",
-                         dispersion = FALSE,
-                         verbose = TRUE,
-                         glmmtmb_zeroinf = FALSE, # needed for edge cases
-                         brms_custom_name = NULL, # needed for edge cases
-                         ...) {
+.retrieve_model_info <- function(x,
+                                 fitfam = "gaussian",
+                                 zero.inf = FALSE,
+                                 hurdle = FALSE,
+                                 logit.link = FALSE,
+                                 multi.var = FALSE,
+                                 link.fun = "identity",
+                                 dispersion = FALSE,
+                                 verbose = TRUE,
+                                 glmmtmb_zeroinf = FALSE, # needed for edge cases
+                                 brms_custom_name = NULL, # needed for edge cases
+                                 ...) {
   dots <- list(...)
   if (isTRUE(dots$return_family_only)) {
     return(list(family = fitfam, link_function = link.fun))
@@ -421,6 +421,7 @@
     is_meta = is_meta,
     is_wiener = inherits(x, "brmsfit") && fitfam == "wiener",
     is_rtchoice = inherits(x, "brmsfit") && fitfam == "custom" && identical(brms_custom_name, "lnr"),
+    is_mixture = fitfam == "mixture",
     link_function = link.fun,
     family = fitfam,
     n_obs = n_obs(x),